+++ tests/test_utils.py
@@ -8,7 +8,8 @@ from ravenpy.utilities.testdata import get_local_testdata
 
 
 class TestOperations:
-    utils = pytest.importorskip("ravenpy.utils")
+    analysis = pytest.importorskip("ravenpy.utilities.analysis")
+    io = pytest.importorskip("ravenpy.utilities.io")
 
     zipped_file = get_local_testdata("polygons/mars.zip")
     non_zipped_file = get_local_testdata("polygons/mars.geojson")
@@ -19,40 +20,30 @@ class TestOperations:
         eastern_angles = np.arange(45, 125, 1.25)
         southwest_angles = np.array([181, 182.25, 183.5, 222])
 
-        assert self.utils.circular_mean_aspect(northern_angles) == 360
+        assert self.analysis.circular_mean_aspect(northern_angles) == 360
         np.testing.assert_almost_equal(
-            self.utils.circular_mean_aspect(slight_northeast_angles), 0, decimal=3
+            self.analysis.circular_mean_aspect(slight_northeast_angles), 0, decimal=3
         )
-        assert self.utils.circular_mean_aspect(eastern_angles) == 84.375
+        assert self.analysis.circular_mean_aspect(eastern_angles) == 84.375
         np.testing.assert_almost_equal(
-            self.utils.circular_mean_aspect(southwest_angles), 191.88055987
+            self.analysis.circular_mean_aspect(southwest_angles), 191.88055987
         )
 
-    def test_parse_lonlat(self):
-        lonlats = ["123,345", "122233344.11111 34554554.2", "1111,2222"]
-        for ll in lonlats:
-            lonlat = self.utils.parse_lonlat(ll)
-            assert len(lonlat) == 2
-            assert isinstance(lonlat[0], float)
-            assert isinstance(lonlat[1], float)
-
-        with pytest.raises(Exception):
-            self.utils.parse_lonlat("This isn't a number, 333.444")
 
     def test_address_append(self):
         non_existing_tarred_file = "polygons.tar"
 
-        assert "zip://" in self.utils.address_append(self.zipped_file)
-        assert "tar://" in self.utils.address_append(non_existing_tarred_file)
-        assert not self.utils.address_append(self.non_zipped_file).startswith(
+        assert "zip://" in self.io.address_append(self.zipped_file)
+        assert "tar://" in self.io.address_append(non_existing_tarred_file)
+        assert not self.io.address_append(self.non_zipped_file).startswith(
             ("zip://", "tar://")
         )
 
     def test_archive_sniffer(self):
-        probable_shp = self.utils.archive_sniffer(self.zipped_file)
+        probable_shp = self.io.archive_sniffer(self.zipped_file)
         assert probable_shp == ["/tmp/mars.shp"]
 
-        probable_shp = self.utils.archive_sniffer(self.zipped_file, working_dir="/tmp")
+        probable_shp = self.io.archive_sniffer(self.zipped_file, working_dir="/tmp")
         assert probable_shp == ["/tmp/mars.shp"]
 
     def test_archive_extract(self):
@@ -62,19 +53,20 @@ class TestOperations:
         files = list()
         with tempfile.TemporaryDirectory() as tdir:
             files.extend(
-                self.utils.generic_extract_archive(self.zipped_file, output_dir=tdir)
+                self.io.generic_extract_archive(self.zipped_file, output_dir=tdir)
             )
             assert len(files) == 5
             for f in files:
                 assert Path(f).exists()
         assert not np.any([Path(f).exists() for f in files])
 
-        files = self.utils.generic_extract_archive(self.zipped_file)
+        files = self.io.generic_extract_archive(self.zipped_file)
         assert np.all([Path(f).exists() for f in files])
 
 
 class TestFileInfoFuncs:
-    utils = pytest.importorskip("ravenpy.utils")
+    checks = pytest.importorskip("ravenpy.utilities.checks")
+    io = pytest.importorskip("ravenpy.utilities.io")
 
     zipped_file = get_local_testdata("polygons/mars.zip")
     geojson_file = get_local_testdata("polygons/mars.geojson")
@@ -85,12 +77,12 @@ class TestFileInfoFuncs:
     non_existing_file = "unreal.zip"
 
     def test_raster_datatype_sniffer(self):
-        datatype = self.utils.raster_datatype_sniffer(self.raster_file)
+        datatype = self.io.raster_datatype_sniffer(self.raster_file)
         assert datatype.lower() == "uint8"
 
     def test_crs_sniffer(self):
-        assert self.utils.crs_sniffer(self.zipped_file) == 4326
-        assert set(self.utils.crs_sniffer(self.geojson_file, self.raster_file)) == {
+        assert self.io.crs_sniffer(self.zipped_file) == 4326
+        assert set(self.io.crs_sniffer(self.geojson_file, self.raster_file)) == {
             4326
         }
 
@@ -99,25 +91,25 @@ class TestFileInfoFuncs:
         zero = []
         three = [1, Path().root, 2.333]
 
-        assert self.utils.single_file_check(one) == one[0]
+        assert self.checks.single_file_check(one) == one[0]
 
         with pytest.raises(FileNotFoundError):
-            self.utils.single_file_check(zero)
+            self.checks.single_file_check(zero)
 
         with pytest.raises(NotImplementedError):
-            self.utils.single_file_check(three)
+            self.checks.single_file_check(three)
 
     def test_boundary_check(self):
         # NOTE: does not presently accept zipped files.
 
         with pytest.warns(None):
-            self.utils.boundary_check([self.geojson_file, self.raster_file], max_y=80)
+            self.checks.boundary_check([self.geojson_file, self.raster_file], max_y=80)
 
         with pytest.warns(UserWarning):
-            self.utils.boundary_check([self.geojson_file, self.raster_file], max_y=15)
+            self.checks.boundary_check([self.geojson_file, self.raster_file], max_y=15)
 
         with pytest.raises(FileNotFoundError):
-            self.utils.boundary_check([self.non_existing_file])
+            self.checks.boundary_check([self.non_existing_file])
 
     @pytest.mark.skip(reason="Not presently testable")
     def test_multipolygon_check(self):
@@ -125,7 +117,7 @@ class TestFileInfoFuncs:
 
 
 class TestGdalOgrFunctions:
-    utils = pytest.importorskip("ravenpy.utils")
+    analysis = pytest.importorskip("ravenpy.utilities.analysis")
     fiona = pytest.importorskip("fiona")
     sgeo = pytest.importorskip("shapely.geometry")
 
@@ -135,26 +127,26 @@ class TestGdalOgrFunctions:
     )
 
     def test_gdal_aspect_not_projected(self):
-        aspect_grid = self.utils.gdal_aspect_analysis(self.raster_file)
+        aspect_grid = self.analysis.gdal_aspect_analysis(self.raster_file)
         np.testing.assert_almost_equal(
-            self.utils.circular_mean_aspect(aspect_grid), 10.9119033
+            self.analysis.circular_mean_aspect(aspect_grid), 10.9119033
         )
 
         # test with creation of a temporary file
         aspect_tempfile = tempfile.NamedTemporaryFile(
             prefix="aspect_", suffix=".tiff", delete=False
         ).name
-        aspect_grid = self.utils.gdal_aspect_analysis(
+        aspect_grid = self.analysis.gdal_aspect_analysis(
             self.raster_file, set_output=aspect_tempfile
         )
         np.testing.assert_almost_equal(
-            self.utils.circular_mean_aspect(aspect_grid), 10.9119033
+            self.analysis.circular_mean_aspect(aspect_grid), 10.9119033
         )
         assert Path(aspect_tempfile).stat().st_size > 0
 
     # Slope values are high due to data values using Geographic CRS
     def test_gdal_slope_not_projected(self):
-        slope_grid = self.utils.gdal_slope_analysis(self.raster_file)
+        slope_grid = self.analysis.gdal_slope_analysis(self.raster_file)
         np.testing.assert_almost_equal(slope_grid.min(), 0.0)
         np.testing.assert_almost_equal(slope_grid.mean(), 64.4365427)
         np.testing.assert_almost_equal(slope_grid.max(), 89.71747, 5)
@@ -162,7 +154,7 @@ class TestGdalOgrFunctions:
         slope_tempfile = tempfile.NamedTemporaryFile(
             prefix="slope_", suffix=".tiff", delete=False
         ).name
-        slope_grid = self.utils.gdal_slope_analysis(
+        slope_grid = self.analysis.gdal_slope_analysis(
             self.raster_file, set_output=slope_tempfile
         )
         np.testing.assert_almost_equal(slope_grid.mean(), 64.4365427)
@@ -170,7 +162,7 @@ class TestGdalOgrFunctions:
 
     # Slope values are high due to data values using Geographic CRS
     def test_dem_properties(self):
-        dem_properties = self.utils.dem_prop(self.raster_file)
+        dem_properties = self.analysis.dem_prop(self.raster_file)
         np.testing.assert_almost_equal(dem_properties["aspect"], 10.9119033)
         np.testing.assert_almost_equal(dem_properties["elevation"], 79.0341721)
         np.testing.assert_almost_equal(dem_properties["slope"], 64.4365427)
@@ -179,7 +171,7 @@ class TestGdalOgrFunctions:
             feature = next(iter(gj))
             geom = self.sgeo.shape(feature["geometry"])
 
-        region_dem_properties = self.utils.dem_prop(self.raster_file, geom=geom)
+        region_dem_properties = self.analysis.dem_prop(self.raster_file, geom=geom)
         np.testing.assert_almost_equal(region_dem_properties["aspect"], 280.6814208)
         np.testing.assert_almost_equal(region_dem_properties["elevation"], 145.8899082)
         np.testing.assert_almost_equal(region_dem_properties["slope"], 61.2650882)
@@ -193,7 +185,7 @@ class TestGdalOgrFunctions:
             geom_1 = self.sgeo.shape(feature_1["geometry"])
             geom_2 = self.sgeo.shape(feature_2["geometry"])
 
-        geom_1_properties = self.utils.geom_prop(geom_1)
+        geom_1_properties = self.analysis.geom_prop(geom_1)
         np.testing.assert_almost_equal(geom_1_properties["area"], 357.9811899)
         np.testing.assert_almost_equal(
             geom_1_properties["centroid"], (-128.3959836, 19.1572278)
@@ -201,7 +193,7 @@ class TestGdalOgrFunctions:
         np.testing.assert_almost_equal(geom_1_properties["perimeter"], 68.4580077)
         np.testing.assert_almost_equal(geom_1_properties["gravelius"], 1.0206790)
 
-        geom_2_properties = self.utils.geom_prop(geom_2)
+        geom_2_properties = self.analysis.geom_prop(geom_2)
         np.testing.assert_almost_equal(geom_2_properties["area"], 361.5114221)
         np.testing.assert_almost_equal(
             geom_2_properties["centroid"], (-70.2394629, 45.7698029)
@@ -211,7 +203,9 @@ class TestGdalOgrFunctions:
 
 
 class TestGenericGeoOperations:
-    utils = pytest.importorskip("ravenpy.utils")
+    analysis = pytest.importorskip("ravenpy.utilities.analysis")
+    geo = pytest.importorskip("ravenpy.utilities.geo")
+
     fiona = pytest.importorskip("fiona")
     rasterio = pytest.importorskip("rasterio")
     sgeo = pytest.importorskip("shapely.geometry")
@@ -226,7 +220,7 @@ class TestGenericGeoOperations:
         reproj_file = tempfile.NamedTemporaryFile(
             prefix="reproj_", suffix=".geojson", delete=False
         ).name
-        self.utils.generic_vector_reproject(
+        self.geo.generic_vector_reproject(
             self.geojson_file, projected=reproj_file, target_crs="EPSG:3348"
         )
 
@@ -235,7 +229,7 @@ class TestGenericGeoOperations:
             feature = next(iterable)
             geom = self.sgeo.shape(feature["geometry"])
 
-        geom_properties = self.utils.geom_prop(geom)
+        geom_properties = self.analysis.geom_prop(geom)
         np.testing.assert_almost_equal(geom_properties["area"], 6450001762792.884, 3)
         np.testing.assert_almost_equal(
             geom_properties["centroid"], (1645777.7589835, -933242.1203143)
@@ -249,7 +243,7 @@ class TestGenericGeoOperations:
         reproj_file = tempfile.NamedTemporaryFile(
             prefix="reproj_", suffix=".tiff", delete=False
         ).name
-        self.utils.generic_raster_warp(
+        self.geo.generic_raster_warp(
             self.raster_file, output=reproj_file, target_crs="EPSG:3348"
         )
 
@@ -269,10 +263,10 @@ class TestGenericGeoOperations:
         reproj_file = tempfile.NamedTemporaryFile(
             prefix="reproj_", suffix=".tiff", delete=False
         ).name
-        self.utils.generic_raster_warp(
+        self.geo.generic_raster_warp(
             self.raster_file, output=reproj_file, target_crs="EPSG:3348"
         )
-        slope_grid = self.utils.gdal_slope_analysis(reproj_file)
+        slope_grid = self.analysis.gdal_slope_analysis(reproj_file)
 
         np.testing.assert_almost_equal(slope_grid.min(), 0.0)
         np.testing.assert_almost_equal(slope_grid.mean(), 0.0034991)
@@ -282,13 +276,13 @@ class TestGenericGeoOperations:
         reproj_file = tempfile.NamedTemporaryFile(
             prefix="reproj_", suffix=".tiff", delete=False
         ).name
-        self.utils.generic_raster_warp(
+        self.geo.generic_raster_warp(
             self.raster_file, output=reproj_file, target_crs="EPSG:3348"
         )
-        aspect_grid = self.utils.gdal_aspect_analysis(reproj_file)
+        aspect_grid = self.analysis.gdal_aspect_analysis(reproj_file)
 
         np.testing.assert_almost_equal(
-            self.utils.circular_mean_aspect(aspect_grid), 7.7805879
+            self.analysis.circular_mean_aspect(aspect_grid), 7.7805879
         )
 
     def test_raster_clip(self):
@@ -299,7 +293,7 @@ class TestGenericGeoOperations:
         clipped_file = tempfile.NamedTemporaryFile(
             prefix="reproj_", suffix=".tiff", delete=False
         ).name
-        self.utils.generic_raster_clip(self.raster_file, clipped_file, geometry=geom)
+        self.geo.generic_raster_clip(self.raster_file, clipped_file, geometry=geom)
 
         with self.rasterio.open(clipped_file) as gt:
             assert gt.crs.to_epsg() == 4326
@@ -314,7 +308,7 @@ class TestGenericGeoOperations:
             feature = next(iter(gj))
             geom = self.sgeo.shape(feature["geometry"])
 
-        transformed = self.utils.geom_transform(geom, target_crs="EPSG:3348")
+        transformed = self.geo.geom_transform(geom, target_crs="EPSG:3348")
         np.testing.assert_almost_equal(
             transformed.bounds,
             (188140.3820599, -2374936.1363096, 3086554.0207066, 409691.2180337),
@@ -325,20 +319,21 @@ class TestGenericGeoOperations:
 
 
 class TestGIS:
-    utils = pytest.importorskip("ravenpy.utils")
+    checks = pytest.importorskip("ravenpy.utilities.checks")
+    io = pytest.importorskip("ravenpy.utilities.io")
     sgeo = pytest.importorskip("shapely.geometry")
 
     vector_file = get_local_testdata("polygons/mars.geojson")
 
     def test_get_bbox_single(self):
-        w, s, n, e = self.utils.get_bbox(self.vector_file, all_features=False)
+        w, s, n, e = self.io.get_bbox(self.vector_file, all_features=False)
         np.testing.assert_almost_equal(w, -139.8514262)
         np.testing.assert_almost_equal(s, 8.3754794)
         np.testing.assert_almost_equal(n, -117.4753973)
         np.testing.assert_almost_equal(e, 29.6327068)
 
     def test_get_bbox_all(self):
-        w, s, n, e = self.utils.get_bbox(self.vector_file)
+        w, s, n, e = self.io.get_bbox(self.vector_file)
         np.testing.assert_almost_equal(w, -139.8514262)
         np.testing.assert_almost_equal(s, 8.3754794)
         np.testing.assert_almost_equal(n, -38.7397456)
@@ -346,7 +341,7 @@ class TestGIS:
 
     def test_feature_contains(self):
         point = -69.0, 45
-        assert isinstance(self.utils.feature_contains(point, self.vector_file), dict)
+        assert isinstance(self.checks.feature_contains(point, self.vector_file), dict)
         assert isinstance(
-            self.utils.feature_contains(self.sgeo.Point(point), self.vector_file), dict
+            self.checks.feature_contains(self.sgeo.Point(point), self.vector_file), dict
         )
